Moved "store" to separated file
