+++ src/app/layout.tsx
@@ -2,41 +2,15 @@ import { Footer, Navbar, Sidebar } from '@/components';
 import './globals.css';
 import type { Metadata } from 'next';
 import { Plus_Jakarta_Sans } from 'next/font/google';
-import { create } from 'zustand';
-import { devtools } from 'zustand/middleware';
-
-interface Environment {
-  name: string;
-  value: string;
-}
-
-interface EnvironmentState {
-  active: Environment['value'];
-  environments: Environment[];
-}
-
-export const useEnvironmentStore = create<EnvironmentState>()(
-  devtools((set) => ({
-    active: 'none',
-    environments: [{ name: 'No Environment', value: 'none' }],
-    setActive: (value: string) => set(() => ({ active: value })),
-    setEnvironments: (environments: Environment[]) =>
-      set(() => ({ environments: environments }))
-  }))
-);
 
 const plusJakartaSans = Plus_Jakarta_Sans({ subsets: ['latin'] });
 
-export const metadata: Metadata = {
+const metadata: Metadata = {
   title: 'Create Next App',
   description: 'Generated by create next app'
 };
 
-export default function RootLayout({
-  children
-}: {
-  children: React.ReactNode;
-}) {
+function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
       <body
@@ -58,3 +32,5 @@ export default function RootLayout({
     </html>
   );
 }
+
+export { RootLayout as default, metadata };
