+++ src/store/types.ts
@@ -1,0 +1,11 @@
+interface Environment {
+  name: string;
+  value: string;
+}
+
+interface EnvironmentState {
+  active: Environment['value'];
+  environments: Environment[];
+}
+
+export type { Environment, EnvironmentState };
